Complete with getCitiesPopulation

+++ src/main/java/telran/java53/person/dto/CityPopulationDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class CityPopulationDto {
 
 	String city;
-	Integer population;
+	Long population;
 }
